+++ NEWS.md
@@ -21,6 +21,11 @@
         strings as column names. Use `pl$lit()` to recover the old behavior.
     -   `$cum_count()` now starts from 1 instead of 0.
 
+### Deprecations
+
+-   `pl$threadpool_size()` is deprecated and will be removed in 0.15.0. Use 
+    `pl$thread_pool_size()` instead (#784).
+
 ## Polars R Package 0.13.1
 
 ### New features